+++ Pinny Notes/Views/SettingsWindow.xaml.cs
@@ -1,4 +1,6 @@
 ﻿using System.Windows;
+
+using Pinny_Notes.Helpers;
 using Pinny_Notes.ViewModels;
 
 namespace Pinny_Notes.Views;
@@ -31,10 +33,27 @@ public partial class SettingsWindow : Window
             if (Owner == _lastOwner)
                 return;
             _lastOwner = Owner;
-            double centerOwnerX = Owner.Left + Owner.Width / 2;
-            double centerOwnerY = Owner.Top + Owner.Height / 2;
-            Left = centerOwnerX - Width / 2;
-            Top = centerOwnerY - Height / 2;
+
+            Point position = new(
+                (Owner.Left + Owner.Width / 2) - Width / 2,
+                (Owner.Top + Owner.Height / 2) - Height / 2
+            );
+            System.Drawing.Rectangle currentScreenBounds = ScreenHelper.GetCurrentScreenBounds(
+                ScreenHelper.GetWindowHandle(Owner)
+            );
+
+            if (position.X < currentScreenBounds.Left)
+                position.X = currentScreenBounds.Left;
+            else if (position.X + Width > currentScreenBounds.Right)
+                position.X = currentScreenBounds.Right - Width;
+
+            if (position.Y < currentScreenBounds.Top)
+                position.Y = currentScreenBounds.Top;
+            else if (position.Y + Height > currentScreenBounds.Bottom)
+                position.Y = currentScreenBounds.Bottom - Height;
+
+            Left = position.X;
+            Top = position.Y;
         }
     }
 }